+++ src/startgg/types.ts
@@ -1,4 +1,4 @@
-export interface iGame {
+export interface sggGame {
   id: string;
   entrant1Score: number;
   entrant2Score: number;
@@ -19,19 +19,35 @@ export interface iGame {
   }[];
 }
 
-export interface iSet {
+export const gqlGame = `
+id
+entrant1Score
+entrant2Score
+state
+stage {
+  id
+  name
+}
+selections {
+  character {
+    id
+    name
+  }
+  entrant {
+    id
+    name
+  }
+}
+`;
+
+export interface sggSet {
   id: string;
   /* 1=Not started, 2=In Progress, 3=Finished */
   state: number;
-  games: iGame[];
+  games: sggGame[];
   winnerId: string;
   fullRoundText: string;
   identifier: string;
-  phaseGroup: {
-    phase: {
-      name: string;
-    };
-  };
   slots: {
     id: string;
     entrant: {
@@ -41,34 +57,108 @@ export interface iSet {
   }[];
 }
 
-export interface iPhase {
-  id: string;
-  name: string;
-  sets: {
-    pageInfo: {
-      total: number;
-    };
-    nodes: iSet[];
-  };
+export const gqlSet = `
+id
+state
+winnerId
+fullRoundText
+identifier
+games {
+  ${gqlGame}
 }
-
-export interface iEvent {
-  id: string;
-  name: string;
-  phases: iPhase[];
-  entrants: {
-    nodes: {
-      id: string;
-      participants: {
-        id: string;
-        gamerTag: string;
-      }[];
-    }[];
-  };
+slots {
+  id
+  entrant {
+    id
+    name
+  }
 }
+`;
 
-export interface iTournamentPeek {
-  id: string;
+export interface sggTournament {
+  id: number;
   name: string;
   slug: string;
+  url: string;
+  isRegistrationOpen: false;
+  eventRegistrationClosesAt: string;
 }
+
+export const gqlTournament = `
+id
+name
+slug
+url
+isRegistrationOpen
+eventRegistrationClosesAt
+`;
+
+export interface sggImage {
+  id: number;
+  url: string;
+}
+
+export const gqlImage = `
+id
+url
+`;
+
+export interface sggOwner {
+  id: number;
+  slug: string;
+  player: { id: number; gamerTag: string; prefix: string };
+}
+
+export const gqlOwner = `
+id
+slug
+player {
+  id
+  gamerTag
+  prefix
+}
+`;
+
+export interface sggEntrant {
+  id: string;
+  participants: {
+    id: string;
+    gamerTag: string;
+  }[];
+}
+
+export const gqlEntrant = `
+id
+participants {
+  id
+  gamerTag
+}
+`;
+
+export interface sggEvent {
+  id: number;
+  name: string;
+  state: string;
+  slug: string;
+}
+
+export const gqlEvent = `
+id
+name
+state
+slug
+`;
+
+export interface sggPhase {
+  id: number;
+  name: string;
+  state: string;
+  bracketType: string;
+}
+
+export const gqlPhase = `
+id
+name
+state
+bracketType
+`;